Add failing test for diffing prefixed vectors
Vectors that have an eltype other than Float64, Int, Char, String, or Symbol
are displayed with a prefix to the array like `Bool[1, 1]`. The `Test.record`
method relies on parsing the expression that is being tested, and when `Bool[1, 1]`
is parsed, it is a `:ref` `Expr`, not a `:vect` expression. This case is not
covered in the logic for checking whether something can have a `deepdiff` or not,
meaning an array like `Bool[1, 1]` does not get `deepdiff`'d.